Added WishlistServiceTests

+++ src/main/java/ca/mcgill/ecse321_group14/GameShop/service/WishlistService.java
@@ -4,6 +4,7 @@ import ca.mcgill.ecse321_group14.GameShop.model.Customer;
 import ca.mcgill.ecse321_group14.GameShop.model.Game;
 import ca.mcgill.ecse321_group14.GameShop.model.Wishlist;
 import ca.mcgill.ecse321_group14.GameShop.repository.CustomerRepository;
+import ca.mcgill.ecse321_group14.GameShop.repository.GameRepository;
 import ca.mcgill.ecse321_group14.GameShop.repository.WishlistRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,18 +18,40 @@ public class WishlistService {
     WishlistRepository wishlistRepository;
     @Autowired
     private CustomerRepository customerRepository;
+    @Autowired
+    private GameRepository gameRepository;
 
+
+    @Transactional
+    public Wishlist createWishlist(Game game, Customer customer) {
+        Wishlist wishlist = wishlistRepository.findWishlistByKey(new Wishlist.Key(game, customer));
+
+        if (wishlist != null) {
+            throw new IllegalArgumentException("Wishlist already exists!");
+        }
+
+        wishlist = new Wishlist(new Wishlist.Key(game, customer));
+        wishlistRepository.save(wishlist);
+        return wishlist;
+    }
 
     @Transactional
     public Wishlist addGameToWishlist(Game game, Customer customer) {
         Wishlist wishlist = wishlistRepository.findWishlistByKey(new Wishlist.Key(game, customer));
-
-        String error = "";
+        if (game == null){
+            throw new IllegalArgumentException("Game does not exist!");
+        }
+        if (customer == null){
+            throw new IllegalArgumentException("Customer does not exist!");
+        }
         if (!customerRepository.existsById(customer.getId())) {
-            error += "Customer does not exist!";
+            throw new IllegalArgumentException("Customer does not exist!");
+        }
+        if (!gameRepository.existsById(game.getId())){
+            throw new IllegalArgumentException("Game does not exist!");
         }
         if (wishlist == null) {
-            wishlist = new Wishlist(new Wishlist.Key(game, customer));
+            throw new IllegalArgumentException("Wishlist does not exist!");
         }
         Wishlist.Key key = new Wishlist.Key(game, customer);
         wishlistRepository.save(wishlist);
@@ -39,6 +62,12 @@ public class WishlistService {
         Wishlist wishlist = wishlistRepository.findWishlistByKey(new Wishlist.Key(game, customer));
         if (wishlist == null) {
             throw new IllegalArgumentException("Wishlist does not exist!");
+        }
+        if(!(wishlist.getKey().getWish().getName().equals(game.getName()))){
+            throw new IllegalArgumentException("Game is not in the wishlist!");
+        }
+        if (!(wishlist.getKey().getBuyer().getUsername().equals(customer.getUsername()))) {
+            throw new IllegalArgumentException("Customer does not own the wishlist!");
         }
         Wishlist.Key key = new Wishlist.Key(game, customer);
         wishlistRepository.deleteById(key);
